Début rfxcom plugin

+++ rfxcom/ressources/rfxcmd/lib/rfx_command.py
@@ -53,11 +53,18 @@ class Command(object):
 			self.process.communicate()
 			logger.debug("Return code: " + str(self.process.returncode))
 			logger.debug("Thread finished")
+			self.timer.cancel()
 		
 		def timer_callback():
-			logger.debug("Thread timeout")
-			self.process.terminate()
-			logger.debug("Thread terminated")
+			logger.debug("Thread timeout, terminate it")
+			if process.is_alive():
+				try:
+					self.process.terminate()
+				except OSError as error:
+					logger.error("Error: %s " % error)
+				logger.debug("Thread terminated")
+			else:
+				logger.debug("Thread not alive")
 			
 		thread = threading.Thread(target=target)
 		self.timer = threading.Timer(int(timeout), timer_callback)
